+++ api/v1/mdb/mongodb_types.go
@@ -1076,8 +1076,7 @@ type Ldap struct {
 }
 
 type OIDCProviderConfig struct {
-	// Unique label that identifies this configuration. This label is visible to your Ops Manager users and is used when
-	// creating users and roles for authorization. It is case-sensitive and can only contain the following characters:
+	// Unique label that identifies this configuration. It is case-sensitive and can only contain the following characters:
 	//  - alphanumeric characters (combination of a to z and 0 to 9)
 	//  - hyphens (-)
 	//  - underscores (_)
@@ -1085,11 +1084,10 @@ type OIDCProviderConfig struct {
 	// +kubebuilder:validation:Required
 	ConfigurationName string `json:"configurationName"`
 
-	// Issuer value provided by your registered IdP application. Using this URI, MongoDB finds an OpenID Provider
+	// Issuer value provided by your registered IdP application. Using this URI, MongoDB finds an OpenID Connect Provider
 	// Configuration Document, which should be available in the /.wellknown/open-id-configuration endpoint.
-	// For MongoDB 7.0, 7.3, and 8.0+, the combination of issuerURI and audience must be unique across OIDC provider configurations.
+	// For MongoDB 8.0+, the combination of issuerURI and audience must be unique across OIDC provider configurations.
 	// For other MongoDB versions, the issuerURI itself must be unique.
-
 	// +kubebuilder:validation:Required
 	IssuerURI string `json:"issuerURI"`
 
@@ -1110,13 +1108,12 @@ type OIDCProviderConfig struct {
 	UserClaim string `json:"userClaim"`
 
 	// The identifier of the claim that includes the principal's IdP user group membership information.
-	// Accept the default value unless your IdP uses a different claim, or you need a custom claim.
 	// Required when selected GroupMembership as the authorization type, ignored otherwise
 	// +kubebuilder:validation:Optional
 	GroupsClaim *string `json:"groupsClaim"`
 
-	// Configure single-sign-on for human user access to Ops Manager deployments with Workforce Identity Federation.
-	// For programmatic, application access to Ops Manager deployments use Workload Identity Federation.
+	// Configure single-sign-on for human user access to deployments with Workforce Identity Federation.
+	// For programmatic, application access to deployments use Workload Identity Federation.
 	// Only one Workforce Identity Federation IdP can be configured per MongoDB resource
 	// +kubebuilder:validation:Required
 	AuthorizationMethod OIDCAuthorizationMethod `json:"authorizationMethod"`