+++ docker/mongodb-kubernetes-tests/tests/conftest.py
@@ -261,10 +261,7 @@ def intermediate_issuer(cert_manager: str, issuer: str, namespace: str) -> str:
     This fixture creates an intermediate "Issuer" in the testing namespace
     """
     # Create the Certificate for the intermediate CA based on the issuer fixture
-    from kubetester.certs import (
-        Certificate,
-        Issuer,
-    )
+    from kubetester.certs import Certificate, Issuer
 
     intermediate_ca_cert = Certificate(namespace=namespace, name="intermediate-ca-issuer")
     intermediate_ca_cert["spec"] = {
@@ -727,7 +724,7 @@ def multi_cluster_operator_manual_remediation(
 
 
 @fixture(scope="module")
-def multi_cluster_operator_no_clustermongodbroles(
+def multi_cluster_operator_no_cluster_mongodb_roles(
     namespace: str,
     central_cluster_name: str,
     multi_cluster_operator_installation_config: dict[str, str],
@@ -1449,10 +1446,7 @@ def create_issuer(
         else:
             raise e
 
-    from kubetester.certs import (
-        ClusterIssuer,
-        Issuer,
-    )
+    from kubetester.certs import ClusterIssuer, Issuer
 
     # And then creates the Issuer
     if clusterwide: