+++ public/tools/multicluster/pkg/common/common.go
@@ -42,20 +42,21 @@ const (
 
 // Flags holds all the fields provided by the user.
 type Flags struct {
-	MemberClusters              []string
-	MemberClusterApiServerUrls  []string
-	ServiceAccount              string
-	CentralCluster              string
-	MemberClusterNamespace      string
-	CentralClusterNamespace     string
-	Cleanup                     bool
-	ClusterScoped               bool
-	InstallDatabaseRoles        bool
-	CreateTelemetryClusterRoles bool
-	OperatorName                string
-	SourceCluster               string
-	CreateServiceAccountSecrets bool
-	ImagePullSecrets            string
+	MemberClusters                []string
+	MemberClusterApiServerUrls    []string
+	ServiceAccount                string
+	CentralCluster                string
+	MemberClusterNamespace        string
+	CentralClusterNamespace       string
+	Cleanup                       bool
+	ClusterScoped                 bool
+	InstallDatabaseRoles          bool
+	CreateTelemetryClusterRoles   bool
+	CreateMongoDBRolesClusterRole bool
+	OperatorName                  string
+	SourceCluster                 string
+	CreateServiceAccountSecrets   bool
+	ImagePullSecrets              string
 }
 
 const (
@@ -521,10 +522,10 @@ func buildClusterRoleTelemetry() rbacv1.ClusterRole {
 	}
 }
 
-func buildClusterRoleMongoDBRole() rbacv1.ClusterRole {
+func buildClusterRoleMongoDBRole(namespace string) rbacv1.ClusterRole {
 	return rbacv1.ClusterRole{
 		ObjectMeta: metav1.ObjectMeta{
-			Name:   DefaultOperatorName + "-multi-cluster-role-mongodb-roles",
+			Name:   fmt.Sprintf("%s-%s-multi-cluster-mongodb-role", DefaultOperatorName, namespace),
 			Labels: multiClusterLabels(),
 		},
 		Rules: []rbacv1.PolicyRule{
@@ -583,7 +584,7 @@ func buildClusterRoleBinding(clusterRole rbacv1.ClusterRole, serviceAccountName,
 }
 
 // createRoles creates the ServiceAccount and Roles, RoleBindings, ClusterRoles and ClusterRoleBindings required.
-func createRoles(ctx context.Context, c KubeClient, serviceAccountName, serviceAccountNamespace, namespace string, clusterScoped, telemetryClusterRoles bool, clusterType clusterType) error {
+func createRoles(ctx context.Context, c KubeClient, serviceAccountName, serviceAccountNamespace, namespace string, clusterScoped, telemetryClusterRoles bool, mongodbRolesClusterRole bool, clusterType clusterType) error {
 	var err error
 
 	if telemetryClusterRoles {
@@ -605,21 +606,25 @@ func createRoles(ctx context.Context, c KubeClient, serviceAccountName, serviceA
 
 	}
 
-	// Create ClusterRole to access the cluster-scoped resource ClusterMongoDBRole
-	clusterRoleForMongoDBRole := buildClusterRoleMongoDBRole()
-	_, err = c.RbacV1().ClusterRoles().Create(ctx, &clusterRoleForMongoDBRole, metav1.CreateOptions{})
-	if err != nil {
-		if errors.IsAlreadyExists(err) {
-			if _, err := c.RbacV1().ClusterRoles().Update(ctx, &clusterRoleForMongoDBRole, metav1.UpdateOptions{}); err != nil {
-				return xerrors.Errorf("error updating role: %w", err)
+	if clusterType == clusterTypeCentral && mongodbRolesClusterRole {
+		// Create ClusterRole to access the cluster-scoped resource ClusterMongoDBRole
+		clusterRoleForMongoDBRole := buildClusterRoleMongoDBRole(serviceAccountNamespace)
+		_, err = c.RbacV1().ClusterRoles().Create(ctx, &clusterRoleForMongoDBRole, metav1.CreateOptions{})
+		if err != nil {
+			if errors.IsAlreadyExists(err) {
+				if _, err := c.RbacV1().ClusterRoles().Update(ctx, &clusterRoleForMongoDBRole, metav1.UpdateOptions{}); err != nil {
+					return xerrors.Errorf("error updating role: %w", err)
+				}
+			} else {
+				return xerrors.Errorf("error creating cluster role: %w", err)
 			}
-		} else {
-			return xerrors.Errorf("error creating cluster role: %w", err)
 		}
-	}
 
-	if err := createClusterRoleBinding(ctx, c, serviceAccountName, serviceAccountNamespace, DefaultOperatorName+"-cluster-mongodb-role-binding", clusterRoleForMongoDBRole); err != nil {
-		return err
+		if err := createClusterRoleBinding(ctx, c, serviceAccountName, serviceAccountNamespace,
+			fmt.Sprintf("%s-%s-multi-cluster-mongodb-role-binding", DefaultOperatorName, serviceAccountNamespace),
+			clusterRoleForMongoDBRole); err != nil {
+			return err
+		}
 	}
 
 	if !clusterScoped {
@@ -705,14 +710,14 @@ func createOperatorServiceAccountsAndRoles(ctx context.Context, clientMap map[st
 		}
 	}
 
-	if err := createRoles(ctx, centralClusterClient, f.ServiceAccount, f.CentralClusterNamespace, f.CentralClusterNamespace, f.ClusterScoped, f.CreateTelemetryClusterRoles, clusterTypeCentral); err != nil {
+	if err := createRoles(ctx, centralClusterClient, f.ServiceAccount, f.CentralClusterNamespace, f.CentralClusterNamespace, f.ClusterScoped, f.CreateTelemetryClusterRoles, f.CreateMongoDBRolesClusterRole, clusterTypeCentral); err != nil {
 		return err
 	}
 
 	// in case the operator namespace (CentralClusterNamespace) is different from member cluster namespace we need
 	// to provide roles and role binding to the operator's SA in member namespace
 	if f.CentralClusterNamespace != f.MemberClusterNamespace {
-		if err := createRoles(ctx, centralClusterClient, f.ServiceAccount, f.CentralClusterNamespace, f.MemberClusterNamespace, f.ClusterScoped, f.CreateTelemetryClusterRoles, clusterTypeCentral); err != nil {
+		if err := createRoles(ctx, centralClusterClient, f.ServiceAccount, f.CentralClusterNamespace, f.MemberClusterNamespace, f.ClusterScoped, f.CreateTelemetryClusterRoles, f.CreateMongoDBRolesClusterRole, clusterTypeCentral); err != nil {
 			return err
 		}
 	}
@@ -735,10 +740,10 @@ func createOperatorServiceAccountsAndRoles(ctx context.Context, clientMap map[st
 			}
 		}
 
-		if err := createRoles(ctx, memberClusterClient, f.ServiceAccount, f.CentralClusterNamespace, f.MemberClusterNamespace, f.ClusterScoped, f.CreateTelemetryClusterRoles, clusterTypeMember); err != nil {
+		if err := createRoles(ctx, memberClusterClient, f.ServiceAccount, f.CentralClusterNamespace, f.MemberClusterNamespace, f.ClusterScoped, f.CreateTelemetryClusterRoles, f.CreateMongoDBRolesClusterRole, clusterTypeMember); err != nil {
 			return err
 		}
-		if err := createRoles(ctx, memberClusterClient, f.ServiceAccount, f.CentralClusterNamespace, f.CentralClusterNamespace, f.ClusterScoped, f.CreateTelemetryClusterRoles, clusterTypeMember); err != nil {
+		if err := createRoles(ctx, memberClusterClient, f.ServiceAccount, f.CentralClusterNamespace, f.CentralClusterNamespace, f.ClusterScoped, f.CreateTelemetryClusterRoles, f.CreateMongoDBRolesClusterRole, clusterTypeMember); err != nil {
 			return err
 		}
 	}